chore: move RelayReply type to amman-client
